improve loader
