+++ packages/@wix-astro/package.json
@@ -11,6 +11,7 @@
   ],
   "exports": {
     ".": "./dist/index.js",
+    "./loaders": "./dist/loaders/index.js",
     "./entrypoints/server.js": "./dist/entrypoints/server.js",
     "./package.json": "./package.json"
   },
