+++ packages/@wix-astro/src/loaders/index.ts
@@ -1,0 +1,3 @@
+import { wixBlogLoader } from "./blog.js";
+
+export { wixBlogLoader };
